feat: Add contact and display modules
